+++ android/src/main/java/me/yohom/amap_map_fluttify/sub_handler/SubHandler25.java
@@ -129,7 +129,7 @@ public class SubHandler25 {
                 // jsonable arg
                 int var1 = (int) ((Map<String, Object>) args).get("var1");
                 // jsonable arg
-                long var2 = (long) ((Map<String, Object>) args).get("var2");
+                int var2 = (int) ((Map<String, Object>) args).get("var2");
             
                 // create target object
                 com.autonavi.ae.gmap.GLMapState obj = new com.autonavi.ae.gmap.GLMapState(var1, var2);
@@ -148,9 +148,9 @@ public class SubHandler25 {
             
                 // args
                 // jsonable arg
-                long var1 = (long) ((Map<String, Object>) args).get("var1");
+                int var1 = (int) ((Map<String, Object>) args).get("var1");
                 // jsonable arg
-                long var3 = (long) ((Map<String, Object>) args).get("var3");
+                int var3 = (int) ((Map<String, Object>) args).get("var3");
             
                 // create target object
                 com.autonavi.ae.gmap.GLMapState obj = new com.autonavi.ae.gmap.GLMapState(var1, var3);
@@ -3418,7 +3418,7 @@ public class SubHandler25 {
                 // jsonable arg
                 Double var6 = (Double) ((Map<String, Object>) args).get("var6");
                 // jsonable arg
-                long var7 = (long) ((Map<String, Object>) args).get("var7");
+                int var7 = (int) ((Map<String, Object>) args).get("var7");
             
                 // create target object
                 com.amap.api.trace.TraceLocation obj = new com.amap.api.trace.TraceLocation(var1, var3, new Double(var5).floatValue(), new Double(var6).floatValue(), var7);
@@ -4134,7 +4134,7 @@ public class SubHandler25 {
                     // jsonable arg
                     int var1 = (int) ((Map<String, Object>) args).get("var1");
                     // jsonable arg
-                    long var2 = (long) ((Map<String, Object>) args).get("var2");
+                    int var2 = (int) ((Map<String, Object>) args).get("var2");
             
                     // create target object
                     com.autonavi.ae.gmap.GLMapState obj = new com.autonavi.ae.gmap.GLMapState(var1, var2);
@@ -4174,9 +4174,9 @@ public class SubHandler25 {
             
                     // args
                     // jsonable arg
-                    long var1 = (long) ((Map<String, Object>) args).get("var1");
+                    int var1 = (int) ((Map<String, Object>) args).get("var1");
                     // jsonable arg
-                    long var3 = (long) ((Map<String, Object>) args).get("var3");
+                    int var3 = (int) ((Map<String, Object>) args).get("var3");
             
                     // create target object
                     com.autonavi.ae.gmap.GLMapState obj = new com.autonavi.ae.gmap.GLMapState(var1, var3);
